Let the user opt if he wants to use .openarena as his fs_homepath.
At the time I wrote this I wasn't with my PC, so this wasn't tested at all. I will not push it until I do or you do. I'm sorry I've been busy all month with school things.

Also, note that the .patch for changing the homepath of quake3 is unnecessary, it can just be a CVAR, I propose we dispose of it.